[refactor] 게시글 조회 페이지 템플릿으로 변경
