--- src/components/pages/Article/style.scss
@@ -1,8 +1,0 @@
-.body_wrapper {
-    display: flex;
-
-    & main {
-        flex: 1;
-        margin-bottom: auto;
-    }
-}
